Move IRI, Item, Locator, Parser and Rating classes to namespace
