+++ library/SimplePie/Rating.php
@@ -41,86 +41,15 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
-/**
- * Handles `<media:rating>` or `<itunes:explicit>` tags as defined in Media RSS and iTunes RSS respectively
- *
- * Used by {@see SimplePie_Enclosure::get_rating()} and {@see SimplePie_Enclosure::get_ratings()}
- *
- * This class can be overloaded with {@see SimplePie::set_rating_class()}
- *
- * @package SimplePie
- * @subpackage API
- */
-class SimplePie_Rating
-{
-	/**
-	 * Rating scheme
-	 *
-	 * @var string
-	 * @see get_scheme()
-	 */
-	var $scheme;
+use SimplePie\Rating;
 
-	/**
-	 * Rating value
-	 *
-	 * @var string
-	 * @see get_value()
-	 */
-	var $value;
+class_exists('SimplePie\Rating');
 
-	/**
-	 * Constructor, used to input the data
-	 *
-	 * For documentation on all the parameters, see the corresponding
-	 * properties and their accessors
-	 */
-	public function __construct($scheme = null, $value = null)
+// @trigger_error(sprintf('Using the "SimplePie_Rating" class is deprecated since SimplePie version 1.x, use "SimplePie\Rating" instead.'), \E_USER_DEPRECATED);
+
+if (\false) {
+	/** @deprecated since SimplePie 1.x, use "SimplePie\Rating" instead */
+	class SimplePie_Rating extends Rating
 	{
-		$this->scheme = $scheme;
-		$this->value = $value;
-	}
-
-	/**
-	 * String-ified version
-	 *
-	 * @return string
-	 */
-	public function __toString()
-	{
-		// There is no $this->data here
-		return md5(serialize($this));
-	}
-
-	/**
-	 * Get the organizational scheme for the rating
-	 *
-	 * @return string|null
-	 */
-	public function get_scheme()
-	{
-		if ($this->scheme !== null)
-		{
-			return $this->scheme;
-		}
-
-		return null;
-	}
-
-	/**
-	 * Get the value of the rating
-	 *
-	 * @return string|null
-	 */
-	public function get_value()
-	{
-		if ($this->value !== null)
-		{
-			return $this->value;
-		}
-
-		return null;
 	}
 }
-
-class_alias('SimplePie_Rating', 'SimplePie\Rating', false);
